Renamed plugins to dependencies, added ExecutableDependecy class
